Create reusable tablable container component.
