+++ src/components/utility/index.ts
@@ -1,0 +1,1 @@
+export * from "./tabLabelContainer";
