+++ src/components/utility/tabLabelContainer/index.ts
@@ -1,0 +1,1 @@
+export { TabLabelContainer } from "./TabLabelContainer";
