+++ src/locale/en.json
@@ -105,9 +105,7 @@
             "commentLabelMarker": "✔ Your latest comment",
             "commentsListTitle": "View Comments",
             "commentsCreatorTitle": "Create Comment",
-            "viewCommentInfo": "View comments",
             "hideCommentInfo": "Hide comments",
-            "createCommentInfo": "Create comment",
             "updateCommentInfo": "Update comment",
             "deleteCommentInfo": "Delete comment",
             "submitCommentInfo": "Submit comment",
