+++ src/locale/sv.json
@@ -105,9 +105,7 @@
             "commentLabelMarker": "✔ Din senaste kommentar",
             "commentsListTitle": "Läs Kommentarer",
             "commentsCreatorTitle": "Skapa Kommentar",
-            "viewCommentInfo": "Se kommentarer",
             "hideCommentInfo": "Dölj kommentarer",
-            "createCommentInfo": "Skapa kommentar",
             "updateCommentInfo": "Uppdatera kommentar",
             "deleteCommentInfo": "Radera kommentar",
             "submitCommentInfo": "Skapa kommentar",
